# Add a flag to vstream to exclude keyspace from table name