--- a/go/vt/proto/vtgate/vtgate_vtproto.pb.go
+++ b/go/vt/proto/vtgate/vtgate_vtproto.pb.go
@@ -435,6 +435,7 @@ func (m *VStreamFlags) CloneVT() *VStreamFlags {
 	r.TabletOrder = m.TabletOrder
 	r.StreamKeyspaceHeartbeats = m.StreamKeyspaceHeartbeats
 	r.IncludeReshardJournalEvents = m.IncludeReshardJournalEvents
+	r.ExcludeKeyspaceFromTableName = m.ExcludeKeyspaceFromTableName
 	if rhs := m.TablesToCopy; rhs != nil {
 		tmpContainer := make([]string, len(rhs))
 		copy(tmpContainer, rhs)
@@ -1846,6 +1847,16 @@ func (m *VStreamFlags) MarshalToSizedBufferVT(dAtA []byte) (int, error) {
 		i -= len(m.unknownFields)
 		copy(dAtA[i:], m.unknownFields)
 	}
+	if m.ExcludeKeyspaceFromTableName {
+		i--
+		if m.ExcludeKeyspaceFromTableName {
+			dAtA[i] = 1
+		} else {
+			dAtA[i] = 0
+		}
+		i--
+		dAtA[i] = 0x50
+	}
 	if len(m.TablesToCopy) > 0 {
 		for iNdEx := len(m.TablesToCopy) - 1; iNdEx >= 0; iNdEx-- {
 			i -= len(m.TablesToCopy[iNdEx])
@@ -2779,6 +2790,9 @@ func (m *VStreamFlags) SizeVT() (n int) {
 			l = len(s)
 			n += 1 + l + protohelpers.SizeOfVarint(uint64(l))
 		}
+	}
+	if m.ExcludeKeyspaceFromTableName {
+		n += 2
 	}
 	n += len(m.unknownFields)
 	return n
@@ -6484,6 +6498,26 @@ func (m *VStreamFlags) UnmarshalVT(dAtA []byte) error {
 			}
 			m.TablesToCopy = append(m.TablesToCopy, string(dAtA[iNdEx:postIndex]))
 			iNdEx = postIndex
+		case 10:
+			if wireType != 0 {
+				return fmt.Errorf("proto: wrong wireType = %d for field ExcludeKeyspaceFromTableName", wireType)
+			}
+			var v int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return protohelpers.ErrIntOverflow
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				v |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			m.ExcludeKeyspaceFromTableName = bool(v != 0)
 		default:
 			iNdEx = preIndex
 			skippy, err := protohelpers.Skip(dAtA[iNdEx:])